+++ mne/decoding/tests/test_transformer.py
@@ -131,32 +131,26 @@ def test_epochs_vectorizer():
     with warnings.catch_warnings(record=True):
         epochs = Epochs(raw, events, event_id, tmin, tmax, picks=picks,
                         baseline=(None, 0), preload=True)
-    epochs_data = epochs.get_data()
     vector = EpochsVectorizer(epochs.info)
-    y = epochs.events[:, -1]
-    X = vector.fit_transform(epochs_data, y)
+    epochs_data = epochs.get_data()
+    X, y = vector.fit_transform(epochs)
 
     # Check data dimensions
     assert_true(X.shape[0] == epochs_data.shape[0])
     assert_true(X.shape[1] == epochs_data.shape[1] * epochs_data.shape[2])
 
-    assert_array_equal(vector.fit(epochs_data, y).transform(epochs_data), X)
+    assert_array_equal(vector.fit(epochs).transform(epochs)[0], X)
+    assert_array_equal(vector.fit(epochs).transform(epochs)[1],
+                       epochs.events[:, -1])
 
     # Check if data is preserved
     n_times = epochs_data.shape[2]
     assert_array_equal(epochs_data[0, 0, 0:n_times], X[0, 0:n_times])
 
     # Check inverse transform
-    Xi = vector.inverse_transform(X, y)
-    assert_true(Xi.shape[0] == epochs_data.shape[0])
-    assert_true(Xi.shape[1] == epochs_data.shape[1])
-    assert_array_equal(epochs_data[0, 0, 0:n_times], Xi[0, 0, 0:n_times])
-
-    # check if inverse transform works with different number of epochs
-    Xi = vector.inverse_transform(epochs_data[0], y)
-    assert_true(Xi.shape[1] == epochs_data.shape[1])
-    assert_true(Xi.shape[2] == epochs_data.shape[2])
+    # epochs_i = vector.inverse_transform(X, y)
+    # assert(epochs_i == epochs)
 
     # Test init exception
-    assert_raises(ValueError, vector.fit, epochs, y)
-    assert_raises(ValueError, vector.transform, epochs, y)
+    assert_raises(ValueError, vector.fit, epochs_data, y)
+    assert_raises(ValueError, vector.transform, epochs_data, y)