+++ doc/whats_new.rst
@@ -29,6 +29,8 @@ API
 
     - Added options for setting data and date formats manually in :func:`mne.io.read_raw_cnt` by `Jaakko Leppakangas`_
 
+    - Added :class:`mne.decoding.XdawnTransformer` transformer to work with two dimensional data array to comply with scikit-learn pipeline. Does mot support events ``overlap`` case by `Asish Panda`_
+
 .. _changes_0_12:
 
 Version 0.12